Replace magic numbers with named constants, use format.Pct consistently
Extract shared layout constants (rightColX, badgeGap, margin, ipY),
use format.Pct for all percentage formatting, and change NetColor/
DiskIOColor to take uint64 matching the collector types.

+++ internal/theme/theme.go
@@ -50,11 +50,11 @@ const (
 func CPUColor(pct float64) uint16  { return ThresholdColor(pct, CPUWarn, CPUCrit) }
 func RAMColor(pct float64) uint16  { return ThresholdColor(pct, RAMWarn, RAMCrit) }
 func DiskColor(pct float64) uint16 { return ThresholdColor(pct, DiskWarn, DiskCrit) }
-func DiskIOColor(v float64) uint16 { return ThresholdColor(v, DiskIOWarn, DiskIOCrit) }
+func DiskIOColor(v uint64) uint16  { return ThresholdColor(float64(v), DiskIOWarn, DiskIOCrit) }
 func APTColor(count int) uint16    { return ThresholdColor(float64(count), APTWarn, APTCrit) }
-func NetColor(v float64, linkSpeedMbps int) uint16 {
+func NetColor(v uint64, linkSpeedMbps int) uint16 {
 	warn, crit := NetThresholds(linkSpeedMbps)
-	return ThresholdColor(v, float64(warn), float64(crit))
+	return ThresholdColor(float64(v), float64(warn), float64(crit))
 }
 
 // ThresholdColor returns a color interpolated across three zones:
